+++ src/constants/numbers.js
@@ -7,5 +7,6 @@
 
 export const HALF = 2;
 export const ONE = 1;
+export const TEN = 10;
 
 export const BLOCK_SIZE = 64;