+++ src/engine/components/mazeGenerator.stories.jsx
@@ -8,7 +8,7 @@ import React from 'react';
 import { storiesOf } from '@kadira/storybook';
 import wait from 'speculation/wait';
 
-import { ONE } from 'constants/numbers';
+import { TEN } from 'constants/numbers';
 import PixiWrapper from 'test/PixiWrapper';
 
 import mazeGenerator from './mazeGenerator';
@@ -18,22 +18,22 @@ import numberGenerator from './numberGenerator';
 const WIDTH = 960;
 const HEIGHT = 448;
 const basicMaze = mazeGenerator();
+const simpleNumberMaze = mazeGenerator();
+const number = numberGenerator(-TEN, MazeData[0]);
 
 storiesOf('engine > components > Maze', module)
   .add('basic Maze', () => (<PixiWrapper component={basicMaze} height={HEIGHT} width={WIDTH} />))
   .add('one moving number Maze', () => {
-    const number = numberGenerator(-ONE);
-
-    basicMaze.addChild(number.view);
-
-    number.updatePosition(MazeData[0]);
+    simpleNumberMaze.addChild(number.view);
 
     (async () => {
       for (let i = 1; i < MazeData.length; i++) {
-        await wait(250); // eslint-disable-line no-magic-numbers
+        await wait(500); // eslint-disable-line no-magic-numbers
         number.updatePosition(MazeData[i]);
       }
+      await wait(2000); // eslint-disable-line no-magic-numbers
+      number.resetPosition();
     })();
 
-    return <PixiWrapper component={basicMaze} height={HEIGHT} width={WIDTH} />;
+    return <PixiWrapper component={simpleNumberMaze} height={HEIGHT} width={WIDTH} />;
   });